resume-analyzer-backend/requirements.txt
fastapi
uvicorn
boto3
python-multipart
pydantic
openai
pdfplumber
python-docx
numpy
python-dotenv
pytest
langchain
pinecone-client
sentence-transformers
transformers
torch
python-dotenv
langchain-community
langchain-pinecone
pymupdf
langchain-huggingface

